[Respects] Add support for replies (#403)
* [Cogs/Respects] Add the reference to the message to which respects are paid

By setting a non-null `reference`, the message could be treated as a reply to the referenced. A null `reference` does not impact the message.
Setting `mention_author` to `False` prevents pinging the referenced message's author.

Reference: https://discordpy.readthedocs.io/en/v1.7.3/api.html#discord.Message.reference

* [Cogs/Respects] Add support for replies

`Respects.payRespects` now queries the last respect (if any, else `None`)  to obtain its reference (i.e., the message to which the respect was paid), and attaches it (the reference) to the new respect message.

`Respects.checkLastRespect` now queries the last respect (if any, else `None`) and checks if the last respect and the current respect (i.e., at the time of invocation of the function) reply to two different messages. If yes, the current respect chain's data (including the user list) gets reset so as to prepare for a new respect chain.

* [Cogs/Respects] Bump logging level to ERROR for HTTPException

HTTP requests should not fail in normal operation. Might signify abnormalities.
- Bump logging level to `ERROR` after catching `discord.HTTPException` (`cogs/respect.py:197`).
- Add `exc_info=True` for extra logging information.

Co-authored-by: Lui <injabie3@gmail.com>

Co-authored-by: Lui <injabie3@gmail.com>