[Respects] Disable errors for non-existent message reference (#417)
* [Cogs/Respects] Prioritize current reference over old reference

Prioritize `ctx.message.reference` over `oldReference` in `payRespects`. Ren should choose the new reference if there is, otherwise should fallback to the reference attached to the previous respect message if there is (and `None` if there is not).

* [Cogs/Respects] Disable errors for non-existent message reference

Set `newReference.fail_if_not_exists` to `False` to avoid Ren raising exceptions when the reference does not exist, in which case the new message would not have a reference anymore.